DPO-864 review fixes, new controller for listing traffic signs and data

+++ src/main/java/fi/livi/digitraffic/tie/data/model/trafficsigns/Device.java
@@ -1,6 +1,5 @@
 package fi.livi.digitraffic.tie.data.model.trafficsigns;
 
-import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -21,10 +20,10 @@ public class Device {
     private String roadAddress;
 
     @Column(name = "etrs_tm35fin_x")
-    private BigDecimal etrsTm35FinX;
+    private Double etrsTm35FinX;
 
     @Column(name = "etrs_tm35fin_y")
-    private BigDecimal etrsTm35FinY;
+    private Double etrsTm35FinY;
 
     public String getId() {
         return id;
@@ -50,19 +49,19 @@ public class Device {
         this.roadAddress = roadAddress;
     }
 
-    public BigDecimal getEtrsTm35FinX() {
+    public Double getEtrsTm35FinX() {
         return etrsTm35FinX;
     }
 
-    public void setEtrsTm35FinX(final BigDecimal etrsTm35FinX) {
+    public void setEtrsTm35FinX(final Double etrsTm35FinX) {
         this.etrsTm35FinX = etrsTm35FinX;
     }
 
-    public BigDecimal getEtrsTm35FinY() {
+    public Double getEtrsTm35FinY() {
         return etrsTm35FinY;
     }
 
-    public void setEtrsTm35FinY(final BigDecimal etrsTm35FinY) {
+    public void setEtrsTm35FinY(final Double etrsTm35FinY) {
         this.etrsTm35FinY = etrsTm35FinY;
     }
